+++ src/aws-provider/aws-credentials/local-aws-profile.ts
@@ -1,11 +1,12 @@
 import { AwsCredentialsType, AwsSdkVersionEnum } from './aws-credentials-type.js';
-import { AwsCredentialsConfig } from '../../types/types.js';
+import {
+  AwsCredentials as AwsCredentialsConfig,
+  LocalAwsProfile as LocalAwsProfileConfig
+} from '../../ops-types.js';
 import { TinyStacksError } from '@tinystacks/ops-core';
 import AWS from 'aws-sdk';
 
-export type LocalAwsProfileConfig = { 
-  profileName: string;
-};
+
 
 class LocalAwsProfile extends AwsCredentialsType implements LocalAwsProfileConfig {
   profileName: string;
