+++ src/index.ts
@@ -7,3 +7,4 @@ export * from './aws-widgets/aws-ecs-deployments.js';
 export * from './aws-widgets/aws-ecs-info.js';
 export * from './utils/utils.js';
 export * from './utils/arn-utils.js';
+export * as OpsTypes from './ops-types.js';
