+++ src/types/types.ts
@@ -1,7 +1,5 @@
-import { AwsAssumedRole, AwsAssumedRoleConfig } from '../aws-provider/aws-credentials/aws-assumed-role.js';
-import { AwsKeys, AwsKeysConfig } from '../aws-provider/aws-credentials/aws-keys.js';
-import { LocalAwsProfile, LocalAwsProfileConfig } from '../aws-provider/aws-credentials/local-aws-profile.js';
-
-export type AwsCredentialsConfig = AwsAssumedRoleConfig | AwsKeysConfig | LocalAwsProfileConfig;
+import { AwsAssumedRole } from '../aws-provider/aws-credentials/aws-assumed-role';
+import { AwsKeys } from '../aws-provider/aws-credentials/aws-keys';
+import { LocalAwsProfile } from '../aws-provider/aws-credentials/local-aws-profile';
 
 export type AwsCredentialsClass = AwsAssumedRole | AwsKeys | LocalAwsProfile;
